Master

See merge request !51

+++ Yamaha/src/CustomCarMsg.cpp
@@ -602,7 +602,7 @@ void CustomCarMsg::updateMoterControlTemp(int value) {
 void CustomCarMsg::updateMoterTemp(int value) {
     MEMBER_PROPERTY_VALUE_CHANGED(moterTemp, value);
 }
-void CustomCarMsg::updateBattery(int value) {
+void CustomCarMsg::updateBattery(float value) {
     MEMBER_PROPERTY_VALUE_CHANGED(battery, value);
 }
 void CustomCarMsg::updateSpn(int value) {
